feat(路径管理): 支持无限路径长度并优化内存管理

+++ legkilo/src/interface/ros2/ros_interface.h
@@ -125,8 +125,8 @@ class RosInterface : public rclcpp::Node {
     double init_time_ = 0.1;
     bool init_flag_ = true;
     
-    // 限制 path 历史长度，避免长时间运行时 path_world_.poses 线性增长占用内存。
-    size_t path_max_size_ = 1000;
+    // path_max_size_ 支持 -1 表示不限制 path 长度，其余非法非正值会在配置读取时回退到 1。
+    int path_max_size_ = 1000;
 
     // 修改说明：odom_only 模式下不再依赖接口层缓存原始点云，保留历史成员声明注释以避免直接删除代码痕迹。
     // CloudPtr cloud_raw_;
